Check AssImp shading model and use it to determine when to use Phong instead of PBR
There are other AssImp shading models, but some of them aren't things the VSG's got a built-in shader for, and there's some dumbness with the available values as the list of options was originally taken from Blender back when Blender's shading models were total nonsense (e.g. Cook-Torrance was actually Blinn-Phong, Blinn-Phong had some parts of Cook-Torrance added etc.).
I've therefore ignored lots of them.

Previously, the value was put into a variable, but then never read, and that only happened on the Phong path, which was more-or-less unreachable as most assimp format loaders set AI_MATKEY_COLOR_SPECULAR for the Phong specular colour, so it wasn't a good signal to use to decide that PBR shaders should be used.