Apply shininess strength to correct parameter
AI_MATKEY_SHININESS_STRENGTH is meant to be a multiplier for the specular colour, not the shininess.
When it's 0.5, specularity should be half as bright, not half as shiny.

I discovered this problem when I realised an FBX file with a 0.25 SpecularFactor field had the unmodified (except for sRGB-to-linear conversion) SpecularColor value in its material uniform buffer.
AssImp automatically applies other factors to their colours for FBX files (e.g. multiplying DiffuseColor by DiffuseFactor before exposing it to us as AI_MATKEY_COLOR_DIFFUSE), but uses AI_MATKEY_SHININESS_STRENGTH for SpecularFactor and exposes SpecularColor directly as AI_MATKEY_COLOR_SPECULAR.

I've verified that AssImp is consistently using AI_MATKEY_SHININESS_STRENGTH to attenuate the specular colour rather than the shininess, so it's just a misleading name rather than its FBX loader being inconsistent with the others.

There's a problem that leaves me not fully satisfied with this solution, though.
For other color/factor combinations, as AssImp is multiplying them in for us, the multiplication's happening in sRGB space, and we're converting to linear RGB afterwards.
The current getColor implementation does the conversion all in one, so we can't slip an extra multiplication in the middle without modifying it.

In a sane world, this would be a good thing, as it doesn't make sense to multiply an sRGB colour by a constant, but as it seems most FBX tools are pre-gamma-correctness or attempting to be compatible with older tools, we're probably doing something inconsistent with most assets we're likely to encounter.