Fix build
3d-tiles.h isn't a file, but 3DTiles.h is.

ushort isn't a type (at least with the headers I've got), but uint16_t is.

SceneGraphBuilder depends on the definition of OperationThreads being available.
I fixed this by including the relevant header in the .h, but it might be preferable to move the definitions of implicit functions into the .cpp.

BatchTable is a struct, so the forward declaration generates a warning if it says it's a class.