Initial attempt at faster SDF generation
The implementation's based on *A Fast Algorithm for Computing the Closest Point
and Distance Transform* by Sean Mauch.

At the moment, it's slower than the previous implementation as I cut some corners to get something that works.
The two key things from this algorithm are that it only computes the distance from edges for texels that might be closest to that edge, and that it does in-polygon tests line-by-line, so it can be done as scanline rasterisation, which can be faster than a from-scratch check for each texel.

I've not used scanline rasterisation, which is a big part of why this is slow, and intend to do that in a later commit.
Also, I've not profiled to see which bits of my new maths are slow and could be rearranged.